requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
fastapi==0.86.0
idna==3.4
numpy==1.23.4
pandas==1.5.1
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.20.4
typing-extensions==4.4.0
urllib3==1.26.12
